Hotfix: Fix failing to fetch metadata and streams due to API restrictions
