+++ app/routes/__init__.py
@@ -1,5 +1,5 @@
 from app.api.mal import MyAnimeListAPI
 
 mal_client = MyAnimeListAPI()
-MAL_ID_PREFIX = "mal"
+MAL_ID_PREFIX = "mal:"
 IMDB_ID_PREFIX = "tt"
